ETL-Project/requirements.txt
numpy == 1.21.5
pandas == 1.5.2
xgboost == 1.7.1
sklearn.model_selection 
train_test_split
sklearn.linear_model 
LinearRegression
sklearn.feature_selection 
LogisticRegression
XGBRegressor
sklearn.ensemble 
RandomForestRegressor
